Sigmoid and Tanh post-operation support for f32 API.
Description:

Implemented sigmoid, tanh as fused post-ops in
aocl_gemm_f32f32f32of32 API's

Sigmoid(x) = 1/1+e^(-x)
Tanh(x) = (1-e^(-2x))/(1+e^(2x))

Updated bench_lpgemm to recognize sigmod, tanh
as options for post-ops from bench_input and verified.

AMD-Internal: [SWLCSG-3178]

+++ bench/bench_aocl_gemm/bench_lpgemm_helpers.h
@@ -273,7 +273,7 @@ static inline ACCUM_type TANH_post_op_ ## BLAS_SFX \
        ACCUM_type temp_accum \
      ) \
 { \
-    double tanh_reference = tanhf( ( double )temp_accum ); \
+    float tanh_reference = tanhf( ( double )temp_accum ); \
     temp_accum = round( tanh_reference ); \
     return temp_accum; \
 } \
